Just Works Update
Doesn't require program restart to create the database on first run, works out of the box now by setting locations to the current directory and has new autodetection features to the FFMPEG and Directory entry windows. Changes the database extension to a .db file, functionality is still the same (FlatFile).