QOL features added
Removed a bit of useless code, cleaned up some comments. Added verification of core files (warns on install and on each launch if there is an error) and checks the database for corruption (basic checks). Retries execution of youtube-dl if failed, and allows user to read the log after execution (enter to continue). Added basic input validation to all inputs and advanced input validation to a few others.